Refactor 'todo_list' app into 'todo_lists'
