--- todo_list/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TodoListConfig(AppConfig):
-    name = "todo_list"
