+++ todo_lists/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TodoListsConfig(AppConfig):
+    name = "todo_lists"
