+++ todo_lists/tests.py
@@ -5,14 +5,14 @@ from rest_framework import status
 from rest_framework.authtoken.models import Token
 from model_bakery import baker
 
-from todo_list.models import TodoList
+from todo_lists.models import TodoList
 
 User = get_user_model()
 
 
 class TodoListTests(TestCase):
     def test_list_todo_lists_requires_authorization(self):
-        url = reverse("todo_list:todo_list")
+        url = reverse("todo_lists:todo_list")
         response = self.client.get(path=url)
         self.assertEqual(status.HTTP_401_UNAUTHORIZED, response.status_code)
 
@@ -32,7 +32,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Client asks for todo lists, none should be available
-        url = reverse("todo_list:todo_list")
+        url = reverse("todo_lists:todo_list")
         response = self.client.get(path=url, **headers)
         self.assertEqual(status.HTTP_200_OK, response.status_code)
         self.assertEqual(0, len(response.data))
@@ -52,7 +52,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Client asks for todo lists successfully
-        url = reverse("todo_list:todo_list")
+        url = reverse("todo_lists:todo_list")
         response = self.client.get(path=url, **headers)
         self.assertEqual(status.HTTP_200_OK, response.status_code)
         self.assertEqual(1, len(response.data))
@@ -70,7 +70,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Attempt to create list, should fail with 400
-        url = reverse("todo_list:todo_list")
+        url = reverse("todo_lists:todo_list")
         response = self.client.post(
             path=url, content_type="application/json", **headers
         )
@@ -91,7 +91,7 @@ class TodoListTests(TestCase):
         # Client creates todo list successfully
         self.assertEqual(0, TodoList.objects.count())
         data_sample = {"title": "This is a sample title for a To-Do List"}
-        url = reverse("todo_list:todo_list")
+        url = reverse("todo_lists:todo_list")
         response = self.client.post(
             path=url, content_type="application/json", data=data_sample, **headers
         )
@@ -107,7 +107,7 @@ class TodoListTests(TestCase):
         self.assertEqual(1, TodoList.objects.count())
 
         # Attempt to retrieve todo list, should fail with 401
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.get(path=url, content_type="application/json",)
         self.assertEqual(status.HTTP_401_UNAUTHORIZED, response.status_code)
 
@@ -127,7 +127,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Attempt to retrieve todo list from 'user 1', should fail with 403
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.get(path=url, content_type="application/json", **headers)
         self.assertEqual(status.HTTP_403_FORBIDDEN, response.status_code)
 
@@ -146,7 +146,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Retrieve todo list
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.get(path=url, content_type="application/json", **headers)
         self.assertEqual(status.HTTP_200_OK, response.status_code)
         self.assertEqual(todo_list.title, response.data["title"])
@@ -164,7 +164,7 @@ class TodoListTests(TestCase):
 
         # Attempt to edit todo list title, should fail with 401
         data_sample = {"title": "This is a new title for the To-Do List"}
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.put(
             path=url, content_type="application/json", data=data_sample,
         )
@@ -185,7 +185,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Attempt to edit todo list title, should fail with 400
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.put(path=url, content_type="application/json", **headers)
         self.assertEqual(status.HTTP_400_BAD_REQUEST, response.status_code)
         self.assertEqual("required", response.data["title"][0].code)
@@ -208,7 +208,7 @@ class TodoListTests(TestCase):
 
         # Attempt to edit todo list title as 'user_1', should fail with 403
         data_sample = {"title": "This is a new title for the To-Do List"}
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.put(
             path=url, content_type="application/json", data=data_sample, **headers
         )
@@ -231,7 +231,7 @@ class TodoListTests(TestCase):
 
         # Edit todo list title successfully
         data_sample = {"title": "This is a new title for the To-Do List"}
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.put(
             path=url, content_type="application/json", data=data_sample, **headers
         )
@@ -249,7 +249,7 @@ class TodoListTests(TestCase):
         self.assertEqual(1, TodoList.objects.count())
 
         # Attempt to delete todo list, should fail with 401
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.delete(path=url, content_type="application/json")
         self.assertEqual(status.HTTP_401_UNAUTHORIZED, response.status_code)
 
@@ -269,7 +269,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Attempto to delete todo list, should fail with 403
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.delete(
             path=url, content_type="application/json", **headers
         )
@@ -290,7 +290,7 @@ class TodoListTests(TestCase):
         headers = {"HTTP_AUTHORIZATION": "Token " + token.key}
 
         # Delete todo list successfully
-        url = reverse("todo_list:todo_list_detail", kwargs={"pk": 1})
+        url = reverse("todo_lists:todo_list_detail", kwargs={"pk": 1})
         response = self.client.delete(
             path=url, content_type="application/json", **headers
         )
