+++ todo_lists/urls.py
@@ -3,7 +3,7 @@ from rest_framework import routers
 
 from .views import TodoListAPIView, TodoListDetailAPIView
 
-app_name = "todo_list"
+app_name = "todo_lists"
 urlpatterns = [
     path("", TodoListAPIView.as_view(), name="todo_list"),
     path("<int:pk>/", TodoListDetailAPIView.as_view(), name="todo_list_detail"),
